When the server sends a message, first invoke the message handler with the specified message.  Then invoke the error handler with error #20028 (SYBESMSG).  This gives us the behavior that Sybase's libraries produce.
This change required modification to the API for
tds_connect().  It is now necessary for the caller to allocate a
TDSSOCKET with tds_alloc_socket(), and set its parent with
tds_set_parent().

Fix existing tds_alloc() calls in the pool program.

Fix tds_alloc_config() so it can work when a NULL locale is
provided.