Add MARS support
This patch add support for MARS in libTDS.

In MARS there could be multiple TDSSOCKET attached to a single TDSCONNECTION.

MARS use an additional header and manage a sort of stack with sequence and
windows in order to multiplex the connection.

Packets are now constructed and appended to a list then if data is request
to be send or received from server the first thread that try to do it
allocate (lock) the connection and do the transfer (even for other sessions).
Although not optimal (cause require extra allocation and copies) this is
easier to implement.