Support DYNPROC capability correctly
If DYNPROC is enabled (and only if enabled) we should prepend prepared
statements with "create proc".

It seems that new server require DYNPROC behaviour (having create proc
in dynamic queries) even if client do not set this flag.

However Sybase SQL Anywhere 10 does not like if flag is turned off
and we use "create proc" so set it correctly.

Problem raised and tested with help of by Tony Esposito.