Use array to store list of sockets in a connection
Using array allow to do fast lookup of sid. Also is quite easy as
sessions must be allocated from in sequence from 0.
When sessions are detached (cause a FIN is sent to the server) I use
an invalid pointer to distinguish from fully closed. This cause you
have to wait to reopen the same session if you don't want the server
to close the connection.