Use TDSSOCKET from stmt instead of dbc
Data are attached to current socket, not only to main one.

In odbc_convert_char, called by odbc_tds2sql stmt->tds can be NULL
(cause is called by SQLGetData) so avoid using it.
This is not perfect as error reporting is done on a socket not attached to
the statement.