Avoid to define _atoi64 if not needed, use _vscprintf to implement vasprintf.
atoll is defined as _atoi64 if available so avoid to define function again
removing a warning is linking option changes (in this case dll export
options).

Microsoft crt functions define a _vscprintf function that return the amount
of characters required to format a given string. Use it to implement vasprintf
if not available.