Fix wrong getopt() option strings so that -d arguments are parsed correctly.
The code treats -d as an alias to -D, but the trailing colon is missing in
the getopt string, which means getopt doesn't return the next argument in
optarg and the code segfaults instead.

Of course, as a result this option has never worked, and it's not documented
in usage, so it would be an equally reasonable fix to drop support for -d
as an alias.

Bug-Debian: http://bugs.debian.org/715697
Bug-Debian: http://bugs.debian.org/715698