Emit a TDSECONF error for invalid configuration
so the user knows that they had an invalid configuration value.

FreeTDS historically has ignored invalid configuration values so if the
user had an invalid configuration setting, it wasn't obvious why things
weren't working the way the user expected. This was particularly bad
with the "encryption" setting -- the user could want encryption and
could easily put an invalid value for it and then would not get
encryption and they would not know why. In fact, at one point, the
FreeTDS userguide had a typo that said that the encryption setting is
"required", but if you look in the code, it's actually "require". So
people would put `encryption = required` and then not get encryption and
they wouldn't realize why.

This makes the connection fail loudly so the user realizes their error.