Use misc/types.csv table to detect type flags
Instead of using long macros prepare an array with flags for each type.
The array is generated with a perl script from misc/types.csv and
include/freetds/proto.h.
Added collate column to misc/types.csv.
Now all flags are the same for Sybase and Microsoft (with a small trick
excluding some Sybase types not implemented), in the future is possible
that we need to know the protocol version (or find another way like use
different numbers).
is_nullable_type was a bit tricky as actually must reflect the BCP types
that in TDS5 are not put into the fixed row data. So for instance SYBCHAR
and SYBNUMERIC are not nullable.
An array of names is also produced (currently commented out).

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>