Make fisql exit with failure when it fails.
It seems that the Sybase flavor of isql does not do this, but the
Microsoft flavor of isql and its descendants do.  But it's gosh
darn convenient when running fisql from a script to be able to
check whether it succeeded, and it isn't going to cause any trouble
for the interactive user, who is highly unlikely to be looking at
the exit status after running fisql.