Support for no terminator in bcp format file
bcp_colfmt requires a null pointer when specifying no terminator, and
currently barfs when zero length is supplied with non-null pointer
(as is the case when specifying an empty string for terminator in
the format file).

When specifying a prefix length and no terminator for all columns in
the format file, freebcp will not perform any fseek on the input
stream, allowing the use of named pipes.

This patch came from Sourceforge.net, patch #128