Handle empty char/varchar in bcp round trip.
Since e30a807 we're not tracking lengths column by column when we
bulk out, which, in the case of of empty strings, causes us to pad
out to the full 256 bytes of the buffer.  If the declared size of
the char or varchar column is less than that length, bulking back
in fails with an overflow error.

The fix for this wasn't simply to avoid padding as it turns out
we weren't really handling empty strings correctly, or at least
weren't doing what the Microsoft bcp client does, which in the
absence of documentation will have to suffice.

What Microsoft bcp does (and what FreeTDS's bcp does after this
commit) is output a single ASCII NUL byte for an empty column on
output, and then on input, handle that NUL character and convert
it back to an empty string, allowing a successful round trip.

Note that this is different from (but easily confused with) the
handling of database NULL, which produces an empty string in the
output file.