Handle empty fields correctly for all types
Instead of testing single types use macro to test for all possible
types, is_variable_type handle all types having a length (which is
different from having a precision).

Also use NUL charater only if we convert to ASCII type, not for instance
if we are using binary BCP format.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>