Fix code copying partial conversion in ODBC
Function to convert characters in ODBC (odbc_set_string_flag) convert
all string discarding part of the string that is not able to write to
output buffer (if this buffer is too small).
On some conditions (depending on encoding) code should copy some bytes
from the discard buffer if there are still some space on output.
The computation of bytes was wrong causing very likely crashes.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>