Disable double conversion and swap bytes.
Code is quite old and is complicated and a bit broken too. Also the
implementation require different layers of conversion which are not
surely the fastest way to do it.

This change proposal was already posted some time ago on ML and raise
no complaints.

Disable it, users will have to use a proper iconv implementation.

Probably HP-UX and Tru64 are affected by this change.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>