Add -E option to datacopy to keep identity values
Identity columns are often used as artificial keys and as such there
are often child tables which make reference to the identity values
from their parents.

Having the ability to keep these values is therefore quite important
as the default behaviour of handing out new values would break and
relationships with child tables.

Signed-off-by: Matthew Green <matthew.green@datamartcomputing.com>
Signed-off-by: Frediano Ziglio <freddy77@gmail.com>