MARS packet fixes and checks
Check correctly data already received in the packet, packet->len is not
the received length but the size wanted on the packet, use
conn->recv_pos instead. If packet was broken in multiple section was
possible to detect a wrong corruption closing the connection.
Also check if packet length is not too big meaning server is trying to
allocate too much data on client. This can happen only on malicious
server or some nasty corruption.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>