Include malloc.h to make alloca work with some Visual Studio version
As reported by Ramiro Morales
(cfr "alloca() usage, building master pre-0.95 on Windows" on ML)
if alloca is called without including this header in Microsoft Visual
C++ Compiler Package for Python 2.7 the linking fails with some missing
symbols.
As this header is not standard (but seems available on all Windows
compiler) include it conditionally.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>