Fix getting TDS protocol version on Windows
On Windows the DM calls SQLGetInfoW even for SQLGetInfo. If the return
is a string not detected by the DM client get confused by the return
of a wide string instead of a multiple byte one.
Change the function to return an integer instead of a string.
This way it works as the interger is encoded in the same way using
SQLGetInfoA or SQLGetInfoW.
This change a bit the ABI from RC1 but was not implemented in any 0.91
(the previous version).

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>