Rename TDS_QUERYING state to TDS_WRITING
TDS_WRITING is more coherent to TDS_READING.
It states that we are writing to server.
This is just a rename to be able to add a new TDS_SENDING state which
like TDS_PENDING allows to distinguish when we are writing to server
or when we are still expecting to send data to server but not
actually doing it (like when using BCP).

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>