Write test to detect problem closing socket with pending data
If a statement with prepared query is closed and there are some
pending data on another statement closing the statement fails.
This is due to the fact that to unprepare a query you need to interact
with the server which connection is busy with another query.
This works with MARS but not without.
Reported by Joel Plotkin (cfr ml "Bug using freetds with perl DBD module")

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>