Add and use new TDS_SENDING state.
This new state is to distinguish when we really are writing to server to
when the server needs more data.
Also allow to better check the state using BCP. To request bulk transfer
to server you just issue some special query which turns the server to
bul mode.
We need to understand when we are switching to bulk mode to avoid libTDS
to send automatically deferred closes on this mode. Previously this
state transition was handled as just going to idle with the risk of
upper layer could send query instead of BCP data.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>