Use ODBC driver register server facility
Instead of registering the ODBC driver manually under Windows use
ActiveX registration to do so. This avoid having to sync installer code
with our driver versions and so on.
Our ODBC driver support ActiveX registration which instead of
registering the ActiveX register the ODBC driver. This trick (the driver
is not an ActiveX) make easier to register the driver in to the system.

Note that installer execute rundll32 instead of calling RegDLL as as
suggested by
http://stackoverflow.com/questions/7618171/not-able-to-register-my-64bit-dll-through-nsis-script
this way make the installer work on both 32 and 64 bit systems.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>