Fix recent make dist failure
Newer autoconf does not like if a source file is included from
different directories making dist target to fail.
So to avoid the problem write some simple C that include the other
file.
Unfortunately this file cannot go to a library won't be
included by the linker as this file contains the Windows
DLL "main" routine.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>