Fix numeric overflow detection in parse_int8
Test was removed time ago but was needed.
Added a test to remember.

The test for res > num is not enough. If MAX is 2^64 and we are
computing num * 10 + digit we mostly can have an overflow if
num > MAX/10. This initially is detected correctly as res will
be mostly 0 and num bigger but as num is bigger we will have something
like res = (num - MAX/10) * 10 so res will be bigger than num at about
(num - MAX/10) * 10 > num -> num - MAX/10 > num/10 ->
num - num/10 > MAX/10 -> num * 10 - num > MAX -> num > MAX/9
(MAX/9 is where the number in the test came from).

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>