Fix typeinfo check for all DM
Make sure if we ask a DATE/TIME/TIMESTAMP using the current date/time
type (which is specific to ODBC version) and we have that type we
return correct row from SQLGetTypeInfo.
Different DMs do different transformation so try to support all types
but return the proper one. For instance if we use ODBC 2 and ask for
SQL_TYPE_TIMESTAMP (which should be used by ODBC 3 applications)
returns SQL_TIMESTAMP (which is the ODBC 2 type).

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>