Try to search some function as symbols in stdio.h
If the function is inline it can happen that trying to detect
that function defining it in a C file and linking the linking
fail as the function is not contained in any library (as inline and
optimised).
So for those function after trying linking try to check into headers
(now stdio.h).

Original patch from Michael Becker

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>