Use only a single query field for TDS_STMT
Query to be executed was stored either in query or in prepared_query
field. The only difference was the kind of query so mainly where the
query was stored was a sort of flag.
Add a flag and just use query field. This avoid some code to check where
the query is stored.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>