Avoid prepared statement results to be carried over different results
Clear current_results if it came from a prepared statement (dynamic).
This can happen if the prepared statement stay alive (referenced outside
the tds socket) as normally is cleared when freed.
Note that cursors results behave in a different way and they have to
remain set even if the socket is in IDLE state.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>