Put all conversion to binary into a single function
Mostly of the conversions to binary are just straightforward memory
copies. Putting all these copies into a single function make most of
the other function easier and reduce code duplication among all of them.
As a consequence of this the resulting binary are quite smaller.
As a result for instance on my machine (x86_64) code is 16% smaller.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>