pool: lot of improvements
Change the way association member <-> user is done. Now a user is
associated with a member till the client connection is closed. This
reduce the number of state required. Previously going to idle the
association was removed but this caused problems with cursors, prepared
statements, transaction and any state staying in the connection.
Put login reply into a new pool_user_send_login_ack. More information
are sent to client and many of them are bound to member ones.
Split pool_user_query in a function to do the association (pool_user_query)
and a function to write data (pool_user_write).

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>