Remove usage of MSG_MORE preferring TCP_CORK
Although TCP_CORK requires slightly more system call (is like being able
to issue a flush of the socket) using it is more complicate requiring to
know in advance if this is the last packet or not.
This can be quite complicated using encryption or MARS.
Using CORK make easier to detect proper point of flushing and also
being more portable (is supported under *BSD) make the code easier to
test.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>