Allow cancel inside message handler
This was a regression if MARS was not enabled.
The socket was marked to be cancelled but the cancellation was not send
but on next query cancel was taken into account and cancel sent to
server causing second query to fail.
This fix TinyTDS problem which issue a dbcancel inside message handler.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>