disable address sanitizer leak detector during configure
Some m4 checks cause leaks making compiled tests generating errors
which are interpreted as errors running code causing functions/libraries
misdetection.
For instance iconv under Ubuntu 15.10 is not detected correctly if
address sanitizer is enabled.
This should not a problem for production as address sanitizer is usually
used as a developer feature.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>