Set socket error if tds_select returns with error
This fix a problem if error was set to EAGAIN already.
Error (errno) was not changed if poll detected a problem on the
connection causing an infinite loop.
This as tds_select use poll and returns error (-1) if the socket
is marked as having error (revents & POLLERR) but errno was not set.

Reported-by: Joshua Lang <joshua.h.lang@gmail.com>
Signed-off-by: Frediano Ziglio <freddy77@gmail.com>