Change optreset test, so that it correctly fails for older mingw versions
Older MingW define optreset to __mingw_optreset however did not
declare optreset so using optreset in source caused a compiler error
as the variable was not defined.
This as configure check declare optreset while other programs don't.
As optreset in MingW should be used only as a compatiblity with BSD
sources avoid to use optreset on such configurations.