move tds_gethostbyname_r to getaddrinfo replacement
This was an old function used to get the address of an host.
Was used before supporting IPv6 and getaddrinfo to have a reentrant
gethostbyname.
Now is used only in systems where getaddrinfo is not present to
implement a getaddrinfo replacement function.
So code is moved in getaddrinfo replacement file and function made
static.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>