Fixed an issue with NTLMV2 being overwritten.
When calling dbsetlbool(login, 1, DBSETNTLMV2); the value would be
overwritten later on during dbopen configuration.

This change adds a use_ntlmv2_specified flag to the login structure
which gets set when dbsetlbool is called with DBSETNTLMV2. The flag is
checked during configuration and if it has been toggled forces
use_ntlmv2 to be as configured.