Add unit tests for batch statements
If you send a batch of statements in a single call to the database, and these statements don't return rows, the dbresults function doesn't process the data correctly. It will loop over all data the first time it is called and return to the calling function that there are no more rows.

Signed-off-by: Jeff Farr jefarr@wayfair.com