configure.ac: support "without" versions of three ./configure flags.
The --with-iodbc, --with-unixodbc, and --with-odbc-nodm flags all use
the standard AC_ARG_WITH macro. The resulting ./configure script
accepts e.g. --without-iodbc to indicate that the user does not want
to enable iODBC, and in place of a path the word "no" is placed into
the $with_iodbc variable. The current configure.ac doesn't handle that
and instead treats "no" as a path. This commit updates configure.ac to
ignore "no" as the value of those three flags.