configure.ac: autodetect iODBC when --with-iodbc=yes.
The default value passed to --with-iodbc (that is, if you don't pass
it an argument) is "yes". Currently, the build system will treat "yes"
as a path and fail to find iODBC there. This commit changes that case,
and uses pkg-config to find the relevant information about iODBC when
--with-iodbc=yes.

This fixes the detection of iODBC on systems where its include files
are located in /usr/include/iodbc (in particular), and addresses
Github issue 142.