configure.ac: autodetect unixODBC when --with-unixodbc=yes.
The default value passed to --with-unixodbc (that is, if you don't
pass it an argument) is "yes". Currently, the build system will treat
"yes" as a path and fail to find unixODBC there. This commit changes
that case, and uses the "odbc_config" executable to find the unixODBC
installation prefix if possible.