doc: Use some SGML entity for iODBC and unixODBC
Make iODBC and unixODBC more consistent in the source.
Also more concise to write.
I split the long lines so to make diff output more helpful in
the future. Currently all sentences in a paragraph are in the
same line which make the diff very unreadable unless you add
some options.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>