doc: Fix userguide generation
Due to recent changes generation were not able to find some
include files.
This as the path were relative to the link in build directory
that was removed.
Instead of creating again the link in the final directory create
them in the doc build directory to the files to include are all
in the same directory making easier to understand.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>