Always use max size for TEXT/NTEXT/IMAGE
The network type information for these types allows to specify a
size. However the definition of the type does not came with a size.
Former implementation does not allows this type to be output and
seems to ignore that length and accept any data even if bigger
(data has also another length field).
Current implementations allows to have BLOBs as output but not
using these types.
All tests passes with these sizes specified so feels safe and at
the same time does not violate the assertion on other types where
the data size has to be maximum the column size.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>