defncopy: Increase sql_text size to make sure is enough for new servers
On MSSQL 2016 onward syscolumns.text field is 8000 characters long.
The sql_text variable that should hold the conversion is 16000 bytes
which are not enough to convert any converted string.
Increase a bit to make sure is big enough.
This is a temporary fix, should be allocated dynamically to make
sure will continue to work.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>