Fix memory leak in tls.c
Each SSL connection that's created would cause a memory leak (11690 bytes
in my testing) because OpenSSL keeps a counter of all certs created with
SSL_get_peer_certificate(). In this case, the certificate was pulled but
never freed, so when SSL_free() was called, OpenSSL saw that it still has
an unfreed certificate and would not free the overall structure.