odbc: Fix SQLEndTran in case auto commit is on
When SQLEndTran is called for a connection with auto commit on
the DM should ignore the call and always return SQL_SUCCESS.
This is documented in Mircosoft documentation too.
This fixes GitHub issue #236 ("Error while trying to execute data
migration") were Django was extracting some data from a database
to migrate data.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>