tds: Avoids connection loop to get stuck
If multiple addresses are present and there's an error on poll
curr_time was not updated causing potentially the loop to be
infinite.
This can happen for instance having multiple addresses to a
not open port. In this case after the first attempt the sockets
are closed and a timeout for a next retry is set however as
the curr_time is never updated code won't never retry new
connections.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>