Do not attempt more connections if the previous one was good
During TDS protocol discovery is possible that the first attempt
fails but we were able to understand that the server did reply
correctly and not refused the version proposed.
This for instance can happen passing a wrong username or password.
This allow the connection to fail faster avoiding to try more
versions than necessary.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>