New CS_DATABASE property
Microsoft Azure does not allow switching databases after the connection
has been made, so we need a way to define the database name prior to the
call to `ct_connect`. Introduce a new, optional CS_DATABASE property
that can be configured with a string value using `ct_con_props`.

I have defined the constant / enum as 9302, as it appears that the other
93XX properties (CS_PORT and CS_CLIENTCHARSET) are also inventions of
FreeTDS.