tds: Fix some issues with server redirection
Server redirection is a feature used by MSSQL, specially Azure
to redirect a connection to another server.
Due to a bug the TLS session was not correctly closed for the
previous attempt causing the second connection to fail is
the first was using TLS (which is the common case for MSSQL).
Another issue was that Azure check the requested "server name"
field in the login to select the databases to provide.
Last was that MARS support was possibly disabled by first attempt
and so MARS support was not enabled for the final connection.

This fixed https://github.com/FreeTDS/freetds/issues/308

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>