odbc: Change SQL_DESC_OCTET_LENGTH value for wide character columns
SQL_DESC_OCTET_LENGTH should return the bytes needed for the buffer
needed to store the default conversion of the column.
For compatibility the default type for wide character columns
is SQL_C_CHAR that is the strings will be converted to multi-byte.
In FreeTDS the charset for multi-byte can be UTF-8 so assuming each
character will take one byte is wrong.
Use "max_bytes_per_char" value.

This fixes issue #305 on GitHub (https://github.com/FreeTDS/freetds/issues/305).

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>