tds: Flush packets if MARS is used
Do not take into account if this is last packet of the query
or not. Is it possible that the next packet of the query won't be
send just after the current one because the sending window does not
allow it. This caused, in case of very long queries, some higher
timeouts causing slowness.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>