tds: Fix return code from tds_iconv in case of errors and same encoding
The error from tds_iconv (like for iconv) in case of errors should
be (size_t) -1, not 0.
In case the source and destination encoding was the same (so
TDS_ENCODING_MEMCPY flag was set) this was not the case.
This could have caused issues if server and client had the same
encoding (which is not much likely).

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>