dblib: Do not report an error if BCP out didn't wrote any row
Fix a regression introduced by "dblib: Set correctly default
BCPMAXERRS and BCPFIRST settings".
This is consistent with MS bcp utility. Sybase one currently reports
an error in case BCPFIRST is set to a value >1 which is a bit inconsistent
with the behaviour if BCPFIRST is 1 (which is not reporting error if
there are no rows in the table).

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>