pool: Close idle connections even if there are no network activity
The idle connections that were to be closed after "max member age"
were not closed if there was no network activity.
This due to the way the main event loop is was written, not accounting
for additional timeouts.
Compute timeout for closing idle connections.
This prevents for instance that momentary heavy load followed by no
load at all, the connections stays active.

This fixes https://github.com/FreeTDS/freetds/issues/347.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>