dblib: Fix column length for terminated fields
Do not ignore length of data if collen < 0.
Do not use strlen if we already have a size.
Threat zero-size as NULL, this is what Sybase does; to pass
an empty field a zero byte or a single space is fine.

bcp2 tests some behaviour of NULL values and bulk copy.

Test based on a test case reported by Bellam R.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>