tds_connect: Leave the original context's error handler alone.
Retire the code that temporarily set the socket's original context's
error handler to NULL during version probing.  The relevant socket
already uses a temporary copy of that context with a suitable custom
error handler, so the only potential usage of this context will be
from other threads that should be able to carry on independently.