Get Windows code page using Windows API
Instead of using setlocale call use GetLocaleInfoA.
This fixes getting Windows code page.
setlocale should return the code page number as first item
but it depends on too many settings and sometimes the string
does not start with the code page number.
Updated ODBC test to avoid not convertible character; note that
now the behaviour is more coherent to Windows driver for this test.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>