Use TDS 7.1 if MARS is not supported
This fix some compatibility with some old MSSQL 2000 server.
We try to negotiate a more recent protocol but if the server does
not support at least TDS 7.2 the DONE token will be shorter causing
FreeTDS to wait for more bytes than sent.
This happens if we give a wrong password logging in to MSSQL 2000
server.
From MSSQL 2005 MARS is returned and we can assume larer DONE
token.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>