Match Sybase ctlib's handling of overlong or malformatted BCP input.
Adjust existing failure modes' boundaries and introduce the
possibility of proceeding with truncated input with or without a
message (with the latter occurring in some cases that require no
conversion).  As for messages, arrange to emit them via _ctclient_msg
rather than _csclient_msg with the same (English) wording and context
as Sybase yields, modulo formal punctuation differences and the usual
discrepancies in layer and origin numbering.

To that end:
* Add internal BLK_CONV_* constants that serve both as (new) client
  message numbers and as possible return values for _cs_convert.
* Cover those numbers in _ct_get_user_api_layer_error.
* Have _cs_convert take the presence of a non-NULL handle (which only
  _blk_get_col_data supplies) to indicate that the caller wants
  bulk-insertion semantics, including in particular the possible use
  of those return values.
* Keep a running count of rows sent for reporting purposes
  (determining the column number to cite by a linear scan of the
  bindinfo's columns array).

Signed-off-by: Aaron M. Ucko <ucko@ncbi.nlm.nih.gov>