unit tests: Formally designate entry points as TEST_MAIN().
Define it in the recently introduced <freetds/utils/test_base.h>
(which now includes <freetds/macros.h> accordingly) as

  int main(int argc TDS_UNUSED, char **argv TDS_UNUSED)

so that main can have the same signature in tests that ignore any
arguments as in tests that consult arguments without yielding any
warnings.  Use this macro for all tests to pave the way for
introducing a central main wrapper around what will then be a
test-specific test_main.

Signed-off-by: Aaron M. Ucko <ucko@ncbi.nlm.nih.gov>