tds: Use inline for tds_socket_init and tds_socket_done
For historical reasons a lot of macros were used instead of
inline functions.
We already use in other places inline functions so it's no
more a portable issue.
The implementation of these function is small and we don't
need additional headers.
This allows reuse in code not linking all libtds library (like
ODBC tests).

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>