odbc: Cache TDS version before initialising parser in tests
For Microsoft driver we need to interrogate the server in order
to understand TDS version.
This was done in odbc_init_parser, however, if we want to reuse
the parser alone it's not a good idea to randomly poke with
the server, so do it explicitly after the connection outside
the parser.
Still the parser could call odbc_tds_version running tds_version_cmp
command.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>