Extend array test to check what happens without status array
Add FLAG_NO_STAT flag which can be used to test the behaviour when
SQL_ATTR_PARAM_STATUS_PTR is not specified.

In this case we can't examine per-row status, but the SQL statement
should still return the expected value, which is different from the one
returned when status array is used in case of error.

This commit is best viewed ignoring whitespace-only changes.

Signed-off-by: Vadim Zeitlin <vz-soci@zeitlins.org>