Swap varint size 4 and 5 to make code more readable
Varint size is, for most types, the size of the length in the protocol.
For instance a size 0 means there's no size associated with the
type while 1 means the size is one byte (so potentially 0-255).
This is not true for size 5 which is a simple 4 bytes length while
type 4 was a bit more complicated (having 2 lengths, one for
text pointer and one for the data).
Swap varint types of 4 and 5 to make the code more readable and
consistent.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>