ctlib: Add a test for data rows locking tables to blk_in test
As other tests for bulk copy skip the test if the setup fails.
This allows to add tests for types not supported by some versions
of the database without having to explicitly test if some types
are supported or not for the specific query.
Add a test with "LOCK DATAROWS" which fails on Microsoft servers.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>