Use CORK flag correctly, do not delay partial packets
Resetting TCP_CORK flag in Linux causes the partial packets
to be flushed however if the queue is full partial packets
are not moved to the queue.
Partial packets are then processed when space is available.
However FreeTDS used to set again the flag causing partial
packets to be delayed.
Remove that delay reducing latency.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>