ODBC MARS: Fix sliding window calculation
The sliding window calculation hardcodes a prefetch size of 4, i.e. the
first fetch will request the server to send 4 packets, which FreeTDS
dequeues into a packet cache. However, on completion of processing of each
one of those packets, it requests the server to send another 4 packets.
This causes the number of packets cached in the client to increase
indefinitely. Changed to base the prefetch window growth on the sequence
number of the packet just processed, not the latest sequence out of all the
cached packets.

Signed-off-by: Matt McNabb <matthew.mcnabb@vmssoftware.com>